guided_tours/src/proxy/requirements.txt
aiohttp==3.5.4
aioredis==1.2.0
async-generator==1.10
async-timeout==3.0.1
asynctest==0.12.2
atomicwrites==1.3.0
attrs==18.2.0
chardet==3.0.4
coverage==4.5.2
docutils==0.14
freezegun==0.3.11
hiredis==1.0.0
idna==2.8
idna-ssl==1.1.0
jmespath==0.9.4
lockfile==0.12.2
more-itertools==6.0.0
multidict==4.5.2
pluggy==0.8.1
py==1.7.0
PyJWT==1.7.1
pytest==4.2.1
pytest-cov==2.6.1
python-daemon==2.2.3
python-dateutil==2.8.0
six==1.12.0
tornado==5.1.1
typing-extensions==3.7.2
urllib3==1.24.1
voluptuous==0.11.5
wrapt==1.11.1
yarl==1.3.0
git+https://github.com/lcgleonard/guided_tours_lib.git@cea5f9860d20a6e4d406fbe2263a0fd5d78c4fa1
